Switch migrate to using Go modules and Go 1.16
This isn't hugely necessary, but we can't ever switch to the upstream
golang-migrate/migrate for various reasons (most notably that its
`schema_migrations` table schema differs from ours, so we can't
actually migrate without dealing with
https://github.com/golang-migrate/migrate/tree/master/database/postgres#upgrading-from-v1. On
top of that, Janus, at the very least, has `CREATE EXTENSION ...`
without an `IF NOT EXISTS`, and things blow up with the upstream
migrate on that too. So...yeah.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>